Add files via upload
Employee Payment for United Industries is an application that enables users to create and store Employee Records. The application also provides facilities of a payroll system that enables the user to generate Payslips. This application is helpful to the department of the organization which maintains data of employees related to an organization.